Add missing license file. Move resolver logic into separate file. Move app sub files into a subfolder.
