+++ LICENSE.md
@@ -1,0 +1,13 @@
+Copyright (c) 2020-2022 ArchiTechAnon
+
+Permission to use, copy, modify, and/or distribute this asset for any
+purpose with or without fee, is hereby granted, provided that the above copyright notice 
+and this permission notice appear in all copies.
+
+THE ASSET IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
+REGARD TO THIS ASSET INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
+AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
+INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
+LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
+OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
+PERFORMANCE OF THIS ASSET.
