+++ app/exceptions.py
@@ -1,0 +1,20 @@
+class Error400BadRequest(Exception):
+    pass
+
+class Error403Forbidden(Exception):
+    pass
+
+class Error403Whitelist(Exception):
+    pass
+
+class Error404NotFound(Exception):
+    pass
+
+class Error408Timeout(Exception):
+    pass
+
+class Error410Gone(Exception):
+    pass
+
+class Error429TooManyRequests(Exception):
+    pass
